FDG-11374 update tests and clean up
